+++ eabm/eabm_package/abaqus/single_element_geometry.py
@@ -12,7 +12,8 @@ def main(output_file, model_name, part_name, width, height):
 
     This script creates a simple Abaqus model with a single rectangle part.
 
-    :param str output_file: The output file for the Abaqus model without extension, e.g. ``output_file``.cae
+    :param str output_file: The output file for the Abaqus model without extension. Will be appended with the required
+        extension, e.g. ``output_file``.cae
     :param str model_name: The name of the Abaqus model
     :param str part_name: The name of the Abaqus part
     :param float width: The rectangle width
@@ -63,7 +64,8 @@ def get_parser():
     parser = argparse.ArgumentParser(description=cli_description,
                                      prog=prog)
     parser.add_argument('-o', '--output-file', type=str, default=default_output_file,
-                        help="The output file for the Abaqus model without extension, e.g. ``output_file``.cae")
+                        help="The output file for the Abaqus model without extension. Will be appended with the " \
+                             "required extension, e.g. ``output_file``.cae")
     parser.add_argument('-m', '--model-name', type=str, default=default_part_name,
                         help="The name of the Abaqus model")
     parser.add_argument('-p', '--part-name', type=str, default=default_part_name,